requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
pydantic
pandas
matplotlib
sqlalchemy